Add weemss page for tickets
