+++ pages/tickets.js
@@ -5,18 +5,6 @@ import Container from '../documentation/Container';
 
 import Tickets from '../patterns/tickets';
 
-const EVENT = {
-  id: 1,
-  name: 'Balkan Ruby',
-  description: 'Lorem, ipsum dolor sit amet consectetur adipisicing elit. Suscipit necessitatibus deserunt minima, reprehenderit odit similique quibusdam libero praesentium cupiditate dignissimos et laboriosam voluptate.',
-  active: true,
-  ticket_types: [
-    {id: 13, name: 'Early Bird', active: false, price: 79.0},
-    {id: 14, name: 'Regular', active: true, price: 99.0},
-    {id: 15, name: 'Workshop', active: true, price: 199.0},
-  ],
-};
-
 export default () => (
   <Layout>
     <Helmet>
@@ -27,7 +15,8 @@ export default () => (
       <div className="container">
         <h1>Tickets</h1>
 
-        <Tickets event={EVENT} />
+        <script src="https://cdn.weemss.com/compiled/js/integration-embed.js?v79.7"></script>
+        <iframe src="https://event.gg/9100/form" id="weemss_integration_9100" frameBorder="0" width="100%" height="100%" scrolling="no"></iframe>
       </div>
     </Container>
   </Layout>
